requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
python-dotenv==1.0.0
SQLAlchemy==2.0.44
boto3
pillow
requests